requirements.txt
googlemaps==4.10.0
googletrans==4.0.0rc1
gTTS==2.2.2
ipinfo==4.4.2
openai==0.27.0
opencv_python==4.7.0.72
Pillow==9.4.0
pillow_heif==0.10.0
pyowm==3.3.0
requests==2.28.2
streamlit_option_menu==0.3.2
whisper==1.1.10
torchvision==0.15.1
torch
audio_recorder_streamlit==0.0.8
numba
torch
tqdm
more-itertools
tiktoken==0.3.3
ffmpeg-python==0.2.0
streamlit-js-eval==0.1.5
altair==4.2.2
numpy==1.23.5
# streamlit==1.8.1
git+https://github.com/openai/whisper.git
google-cloud
google-cloud-vision
tensorflow
